Code improvements and minor refactor
