+++ src/meta/data/RecordAssessmentData.ts
@@ -1,9 +1,8 @@
 import { CountryIso } from '@meta/area'
-import { AssessmentName, CycleName, NodeValue, TableName, VariableName } from '@meta/assessment'
+import { AssessmentName, ColName, CycleName, NodeValue, TableName, VariableName } from '@meta/assessment'
 
-export type ColumnName = string
 export type RecordRowData = Record<VariableName, NodeValue>
-export type RecordColumnData = Record<ColumnName, RecordRowData>
+export type RecordColumnData = Record<ColName, RecordRowData>
 export type RecordTableData = Record<TableName, RecordColumnData>
 export type RecordCountryData = { [key in CountryIso]?: RecordTableData }
 export type RecordCycleData = Record<CycleName, RecordCountryData>
