+++ src/meta/data/recordAssessmentDatas/getAssessmentData.ts
@@ -1,0 +1,5 @@
+import { Props } from './props'
+
+export const getAssessmentData = (props: Pick<Props, 'assessmentName' | 'data'>) => {
+  return props.data?.[props.assessmentName] ?? {}
+}
