+++ src/meta/data/recordAssessmentDatas/getDatum.ts
@@ -1,0 +1,6 @@
+import { getNodeValue } from '@meta/data/recordAssessmentDatas/getNodeValue'
+import { Props } from '@meta/data/recordAssessmentDatas/props'
+
+export const getDatum = (props: Props): string | undefined => {
+  return getNodeValue(props)?.raw
+}
