+++ src/server/controller/cycleData/getBulkDownload/getFRAYearsData.ts
@@ -1,3 +1,5 @@
+import { RecordAssessmentDatas } from '@meta/data'
+
 import { entries } from '@server/controller/cycleData/getBulkDownload/entries/FRAYears'
 import { genders } from '@server/controller/cycleData/getBulkDownload/genders'
 import { getClimaticValue } from '@server/controller/cycleData/getBulkDownload/getClimaticValue'
@@ -9,7 +11,12 @@ import { Props } from './props'
 
 export const getFraYearsData = async (props: Props) => {
   const { assessment, cycle, countries } = props
-  const climaticData = await climaticDomain(props)
+  const _climaticData = await climaticDomain(props)
+  const climaticData = RecordAssessmentDatas.getCycleData({
+    assessmentName: assessment.props.name,
+    cycleName: cycle.name,
+    data: _climaticData,
+  })
   const tableNames = entries.map(({ tableName }) => tableName)
   const tableData = await getData({
     assessment,
@@ -18,7 +25,11 @@ export const getFraYearsData = async (props: Props) => {
     tableNames,
   })
 
-  const data = tableData[assessment.props.name][cycle.name]
+  const data = RecordAssessmentDatas.getCycleData({
+    assessmentName: assessment.props.name,
+    cycleName: cycle.name,
+    data: tableData,
+  })
 
   // Unique years
   const years = getYears({
@@ -34,22 +45,48 @@ export const getFraYearsData = async (props: Props) => {
         iso3: countryIso,
         name: countryIso,
         year,
-        boreal: getClimaticValue('boreal', countryIso, climaticData[assessment.props.name][cycle.name]),
-        temperate: getClimaticValue('temperate', countryIso, climaticData[assessment.props.name][cycle.name]),
-        tropical: getClimaticValue('tropical', countryIso, climaticData[assessment.props.name][cycle.name]),
-        subtropical: getClimaticValue('sub_tropical', countryIso, climaticData[assessment.props.name][cycle.name]),
+        boreal: getClimaticValue('boreal', countryIso, climaticData),
+        temperate: getClimaticValue('temperate', countryIso, climaticData),
+        tropical: getClimaticValue('tropical', countryIso, climaticData),
+        subtropical: getClimaticValue('sub_tropical', countryIso, climaticData),
       }
 
       entries.forEach(({ variables, tableName }) => {
         variables.forEach(({ variableName, csvColumn }) => {
           if (tableName === 'carbonstocksoildepth')
-            base[csvColumn] = data[countryIso][tableName]?.[variableName]?.[variableName]?.raw ?? null
+            base[csvColumn] = RecordAssessmentDatas.getDatum({
+              assessmentName: assessment.props.name,
+              cycleName: cycle.name,
+              data,
+              countryIso,
+              tableName,
+              variableName,
+              colName: variableName,
+            })
           else if (tableName === 'graduationofstudents' || tableName === 'employment') {
             genders.forEach((gender) => {
               base[`${csvColumn}_${gender.csv}`] =
-                data[countryIso][tableName]?.[`${year}_${gender.variable}`]?.[variableName]?.raw ?? null
+                RecordAssessmentDatas.getDatum({
+                  assessmentName: assessment.props.name,
+                  cycleName: cycle.name,
+                  data,
+                  countryIso,
+                  tableName,
+                  variableName,
+                  colName: `${year}_${gender.variable}`,
+                }) ?? null
             })
-          } else base[csvColumn] = data[countryIso][tableName]?.[year]?.[variableName]?.raw ?? null
+          } else
+            base[csvColumn] =
+              RecordAssessmentDatas.getDatum({
+                assessmentName: assessment.props.name,
+                cycleName: cycle.name,
+                data,
+                countryIso,
+                tableName,
+                variableName,
+                colName: year,
+              }) ?? null
         })
       })
